+++ pkg/resources/database/resource_service_redis_test.go
@@ -23,7 +23,7 @@ import (
 	"github.com/exoscale/terraform-provider-exoscale/pkg/testutils"
 )
 
-type TemplateModelGrafana struct {
+type TemplateModelRedis struct {
 	ResourceName string
 
 	Name string
@@ -34,18 +34,19 @@ type TemplateModelGrafana struct {
 	MaintenanceTime       string
 	TerminationProtection bool
 
-	IpFilter        []string
-	GrafanaSettings string
+	IpFilter      []string
+	RedisSettings string
 }
 
-func testResourceGrafana(t *testing.T) {
-	tpl, err := template.ParseFiles("testdata/resource_grafana.tmpl")
+//lint:ignore U1000 redis EOL
+func testResourceRedis(t *testing.T) {
+	tpl, err := template.ParseFiles("testdata/resource_redis.tmpl")
 	if err != nil {
 		t.Fatal(err)
 	}
 
-	fullResourceName := "exoscale_database.test"
-	dataBase := TemplateModelGrafana{
+	fullResourceName := "exoscale_dbaas.test"
+	dataBase := TemplateModelRedis{
 		ResourceName:          "test",
 		Name:                  acctest.RandomWithPrefix(testutils.Prefix),
 		Plan:                  "hobbyist-2",
@@ -57,7 +58,7 @@ func testResourceGrafana(t *testing.T) {
 	dataCreate.MaintenanceDow = "monday"
 	dataCreate.MaintenanceTime = "01:23:00"
 	dataCreate.IpFilter = []string{"1.2.3.4/32"}
-	dataCreate.GrafanaSettings = strconv.Quote(`{"disable_gravatar":true}`)
+	dataCreate.RedisSettings = strconv.Quote(`{"io_threads":1,"lfu_decay_time":1,"lfu_log_factor":10,"maxmemory_policy":"noeviction","notify_keyspace_events":"","persistence":"rdb","ssl":true,"timeout":300}`)
 	buf := &bytes.Buffer{}
 	err = tpl.Execute(buf, &dataCreate)
 	if err != nil {
@@ -69,7 +70,7 @@ func testResourceGrafana(t *testing.T) {
 	dataUpdate.MaintenanceDow = "tuesday"
 	dataUpdate.MaintenanceTime = "02:34:00"
 	dataUpdate.IpFilter = nil
-	dataUpdate.GrafanaSettings = strconv.Quote(`{"allow_embedding":true,"disable_gravatar":true}`)
+	dataUpdate.RedisSettings = strconv.Quote(`{"io_threads":1,"lfu_decay_time":1,"lfu_log_factor":10,"maxmemory_policy":"noeviction","notify_keyspace_events":"","persistence":"rdb","pubsub_client_output_buffer_limit":64,"ssl":true,"timeout":300}`)
 	buf = &bytes.Buffer{}
 	err = tpl.Execute(buf, &dataUpdate)
 	if err != nil {
@@ -79,7 +80,7 @@ func testResourceGrafana(t *testing.T) {
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testutils.AccPreCheck(t) },
-		CheckDestroy:             CheckServiceDestroy("grafana", dataBase.Name),
+		CheckDestroy:             CheckServiceDestroy("redis", dataBase.Name),
 		ProtoV6ProviderFactories: testutils.TestAccProtoV6ProviderFactories,
 		Steps: []resource.TestStep{
 			{
@@ -94,7 +95,7 @@ func testResourceGrafana(t *testing.T) {
 					resource.TestCheckResourceAttrSet(fullResourceName, "ca_certificate"),
 					resource.TestCheckResourceAttrSet(fullResourceName, "updated_at"),
 					func(s *terraform.State) error {
-						err := CheckExistsGrafana(dataBase.Name, &dataCreate)
+						err := CheckExistsRedis(dataBase.Name, &dataCreate)
 						if err != nil {
 							return err
 						}
@@ -108,7 +109,7 @@ func testResourceGrafana(t *testing.T) {
 				Config: configUpdate,
 				Check: resource.ComposeAggregateTestCheckFunc(
 					func(s *terraform.State) error {
-						err := CheckExistsGrafana(dataBase.Name, &dataUpdate)
+						err := CheckExistsRedis(dataBase.Name, &dataUpdate)
 						if err != nil {
 							return err
 						}
@@ -133,7 +134,7 @@ func testResourceGrafana(t *testing.T) {
 	})
 }
 
-func CheckExistsGrafana(name string, data *TemplateModelGrafana) error {
+func CheckExistsRedis(name string, data *TemplateModelRedis) error {
 	client, err := testutils.APIClient()
 	if err != nil {
 		return err
@@ -141,7 +142,7 @@ func CheckExistsGrafana(name string, data *TemplateModelGrafana) error {
 
 	ctx := exoapi.WithEndpoint(context.Background(), exoapi.NewReqEndpoint(testutils.TestEnvironment(), testutils.TestZoneName))
 
-	res, err := client.GetDbaasServiceGrafanaWithResponse(ctx, oapi.DbaasServiceName(name))
+	res, err := client.GetDbaasServiceRedisWithResponse(ctx, oapi.DbaasServiceName(name))
 	if err != nil {
 		return err
 	}
@@ -159,20 +160,20 @@ func CheckExistsGrafana(name string, data *TemplateModelGrafana) error {
 	}
 
 	if !cmp.Equal(data.IpFilter, *service.IpFilter, cmpopts.EquateEmpty()) {
-		return fmt.Errorf("grafana.ip_filter: expected %q, got %q", data.IpFilter, *service.IpFilter)
+		return fmt.Errorf("redis.ip_filter: expected %q, got %q", data.IpFilter, *service.IpFilter)
 	}
 
 	if v := string(service.Maintenance.Dow); data.MaintenanceDow != v {
-		return fmt.Errorf("grafana.maintenance_dow: expected %q, got %q", data.MaintenanceDow, v)
+		return fmt.Errorf("redis.maintenance_dow: expected %q, got %q", data.MaintenanceDow, v)
 	}
 
 	if data.MaintenanceTime != service.Maintenance.Time {
-		return fmt.Errorf("grafana.maintenance_time: expected %q, got %q", data.MaintenanceTime, service.Maintenance.Time)
+		return fmt.Errorf("redis.maintenance_time: expected %q, got %q", data.MaintenanceTime, service.Maintenance.Time)
 	}
 
-	if data.GrafanaSettings != "" {
+	if data.RedisSettings != "" {
 		obj := map[string]interface{}{}
-		s, err := strconv.Unquote(data.GrafanaSettings)
+		s, err := strconv.Unquote(data.RedisSettings)
 		if err != nil {
 			return err
 		}
@@ -182,9 +183,9 @@ func CheckExistsGrafana(name string, data *TemplateModelGrafana) error {
 		}
 		if !cmp.Equal(
 			obj,
-			*service.GrafanaSettings,
+			*service.RedisSettings,
 		) {
-			return fmt.Errorf("grafana.grafana_settings: expected %q, got %q", obj, *service.GrafanaSettings)
+			return fmt.Errorf("redis.redis_settings: expected %q, got %q", obj, *service.RedisSettings)
 		}
 	}